Added includes for previously unincluded files to appease the Strict Mode Gods.
CoC now compiles in strict mode, and all AS has been removed from the FLA.  Well, there may still be some in the Flex components, but We Don't Talk About Flex.  (JK love you Flex.  Modified love.)  For some reason, Flash wasn't seeing the jojoRape() function, even though it was in the same scope as all the other functions in the program, so I added a property to classes.CoC: semiglobalReference.  It's a plain object to which you can attach things like functions or other global properties as a bodge to make things build.  This is a bad thing, but since we're going to restructure everything eventually anyway, I guess it's okay?  Given that by all rights the jojoRape function should work anyway, it's likely due to Flash's compiler being silly.  It has done this to me before and it took forever to fix, and I have no idea what I did.