Overhauling the Creature class
Overhauled the creature class to break it down into separate classes.
Creature is the top class, Monster and Character inherit from it and
Player and NonPlayer inherits from Character. NonPlayer is currently
useless and I have high hopes for it to convert over Followers to
classes.

I've also included something I'm testing out in Cocks, using an enum
substitute that I found online. CockTypes is a basic version of it,
ItemsEnum is something a little more complicated. I think this may
really give us a clear structure to all the types floating around.