End of night commit: Starting to switch to using MainView as the document class.
FLA is now set to generate lib/CoC.swc and this will eventually be used as the view instead of just compiling the FLA directly.  After that is testing with mxmlc to compile CoC.as, then creating a Flash Builder project.  Also, fixed a few bugs to get the swc to build.

NOTE: I have not touched CoC (CS6).fla because I'm using CC, so these two FLAs are out of sync.