End of night commit: Slowly trawling through original view code, ripping it out, and replacing functions with pass throughs to methods on the new View classes.
CoC now has a model property, of type *, because it's a bodge right now and will be used to pass game state through to the View classes.  Eventually, we'll replace this with a real Model class of some sort, maybe with update events and all that fancy shit.

Some day when we've gotten everything better organized, that is.