Moved view classes to their own class tree to prevent possibly overriding the SWC definitions with the bare class definitions non-view-lib source.
I thought I'd read that that would happen at some point, but I can't seem to track down the post.  The main classes shouldn't see the view classes in the source at all, anyway, so oh well.  Also updated the FLA to point to lib-classes instead of classes.  It compiles without errors but that's not saying much.