Big Update adding image support, fixing some save issues and various bugs
Added ImageManager and Image classes to handle adding images to the game. Included is a program called XMLBuilder (source in difProc) that builds an xml file in /img/ of all the images in /img/. Filenames must conform to the ids list in ImageFileKey to appear in game. I've included an example images.xml that will display some existing images for your viewing pleasure. ImageFilesKey is a makeshift dictionary file that just keeps track of all the current id names that I've written in so far. Too late to remove now but I'll add it to the wiki instead.
Fixed some save.as bugs that prevented the buttons from showing.