End of night commit; Must remember that my first goal is to simply separate the UI.
Refactoring can come later, first it is necessary to separate things even if they remain tightly coupled.  So, it's likely I'll be assigning a number of callbacks directly to CoC.mainView to facilitate the original behavior.  This means not doing anything funny with the button execution code, yet, no matter how much I want to.