Wooo boy. This change is mostly around refactoring discriptors in Appearance.
But it also includes changes to how the images work. It now doesn't require regenerating the list, we just need to provide a "master list" of all the refered images. Changes also to CockEnum to remove legacy stuff, which included writing a new method to refer to types by index. Incidently, I think this also fixed a compatiblity issue with Fen's version, because now all saves just use a number, like before. Also includes more images in the game!