Reworked the ImageManager system again, this time to make it work without all the hardships of loading in the xml.
It is also fast, which was a nice bonus. New folder devTools to house the exe that builds the production xml that is packaged into the swf. Not intended to be released, it is only used to build the "master" xml file that includes all currently included images. It should be run whenever you made a new release. Also, it runs from devTools, don't move it because it doesn't have those nice features herp-derp included to protect you from yourself.