Fixed an issue with the tailor flirt.
There was a nonsensical y variable that could point to a nonexistent
cock and the option to set it to your smallest dick if you were too big
was commented out for some reason.

Not sure if you guys commented those out globally, but in some scenes
they're actually set up to handle any size, they just try for certain
range. It isn't always just a failsafe in case a PC with a dick that's
too big somehow glitches in.