Typos and bug fixes
A couple bug fixes here, the first are fixing issue(s) #237 (as well as
#236) and #238. (My apologies if it isn't hyperlinking correctly as my
experience with github is limited if anything) The first wouldn't allow
players to access Rathazul's armor menu. This was caused by a use of AND
operators instead of OR operators in an if statement; this has been
corrected. The second problem stemmed from the inability to access the
menu where you may give Kid A one of your weapons. At first I thought
that the method simply wasn't being called, however my lack of knowledge
in ActionScript was the cause of this mistake. It was actually my
misinterpretation of the coder's intentions. There is an if statement
that was there and inhibiting the call to the method to activate the
menu. I am sure there was a purpose for this if statement, however it is
stopping a feature from existing so I believe it best to be taken out
for now until the coder has decided what to do with it.

As for the typos, they are #240 and #242. The AnsScene.as was quite a
simple fix. The UtraQuest, however, puzzled me at first; I was nor am
able to thing of a better fix than "in seconds". For now I hope that
typo fix is fine, but I'm sure the author(s) can find a better way to
phrase it if they so choose.