Fixes #292
Blanketed a lot of the reminaing instances of " his " in the text with the arian tags, mostly where appropriate. I might have stuck some in scenes where we can implicitly know that Arian will be male, but then, if Arian is male for the scene, the tags should return correctly anyway, right?