Fixes #317
I think this is the correct mechanic; C.Clothes can be "transformed" by applying a modName that gets saved. If modArmorName exists when attempting to get the armor name of the player, modArmorName is returned instead. Removing any armor (I think, I did the clearing in the base Armor class, and I'm not 100% sure all armor is using this? maybe?) with a modArmorName value set will clear the var and revert to using the actual armor name the player is wearing.
This should also automatically fix null values for the modArmorName string in saves, and convert them to an actual empty string.