Working up the new build stuff required to get this AIR version off the floor. This is going to complicate the build process.
Splitting the "Project" files into Standalone and AIR variants with the compiler arguments set to build the two types.
The FlashDevelop AIR version is setup to test the app with adl post-compile too. Need to fiddle with ignore in a minute and tweak the app.xml some more for FD. And then have arguments with the Flash IDE.