Quickfix: Kitsune's Gift post-fight management.
leaveKitsune(talked) was defaulted to true (and called as true to boot) when talking and leaving, but never had a chance to get called as false. Now defaulting to false and calling it (as default, false) from the post-fight Leave button as well. The existing inCombat check should take care of the cleanup as intended. The Talk callback will still call it as true and give the expected result.

Rebranching from master because berserk Holli. Sorry for all the branches.