Fix for Fetish Zelaot armor name change for males
In the Fetish Zealot loss scene the player's armor can possibly be
changed via a modArmorName call. For some reason the male and female
armor name changes were identical. This meant a male player never got
the benefit of a tease attack, as it didn't match what was expected in
Combat.as - changed to match Combat.as