Optimization and minor function call fixes
Only adresses issue
#756

Although it has other optimizations and minor fixes.

Note: Scenes\Places\TelAdre\Katherine.as line 5285 (or 5287 if merged)
There is a hasCock() condition inside the hasCock() condition on line
5283 (5285).
I don't really get what the writer wanted to do there, so for the time
being, I decided to comment it out for the time being.