Anal and vaginal stretching capped at 5
There was inconsistent behaviour when vagina or ass could be stretched
beyond CLOWN_CAR level if cockarea was < capacity, but could not be
stretched when cockArea was >= capacity. This is now fixed by setting
maximum stretch of 5. Also, vagina and ass are recovered if they are
stretched beyond 5 (just in case)