Game over texts fix and Anzu Scene parser experiments
There was a typo in the game over texts code:
`textChoices += 4 + rand(5);` should read `textChoices = 4 + rand(5);` in line 40

In addition to that, I've experimented a little with the parser in AnzuScene.as. Mainly I wanted to replace
`" + player.skin() + "` with `[skin]` in order to avoid conflicts with my BodyParts.UnderBody work, but I didn't want to stop there and played around a little more with the parser.

@kitteh6660: If you don't like that, simple ditch that PR and I'll redo it without the parser experiments and only do the `" + player.skin() + "` to `[skin]` stuff. Generally its just a matter of personal preferrence anyway.