Lethicite Staff Crafting
Added the ability for Rathazul to craft the Lethicite Staff via 5 Lethicite and a Wizard's Staff (which gets imbued with the Lethicite). Likewise, changed the armor button "Craft" since I filed it under there and it wouldn't make sense for a weapon to be under armor. On the plus side, more opportunities for the future.