Update KihaFollower.as
Made changes to where instead of children being referred to as "her" they're either now referred to as "the" or "your" to make it feel less like you're not a parent. Will go back, and change if not acceptable
Made some changes to word flow to better fit sentence structures for interactions with child/children. 
Added in a bit of coding pertaining to an oversight during story time for child/children scenes.