PerkDB 2.0 Phase 1. Next, adding interaction menus to show perk information.
Also. Please make sure all perks use Requirements for checking for player eligibility, now and in the future. Check Page 0 for all perks that don't use this and identify which needs to be reworked.

Perks Should be reworked sooner than later. Preferably, with also an argument for tagging class as well, "Weapon", "Armour", "EnemyOnly", "Player", etc.