more tweaks. Not sure if you still want to include it into the release later on, but easy enough to hide for that release.
Though, if we ever want to release it if that's what we are concerned about... perks will need to all run on requirements. In which case, a rework might be nice, to reduce the number of perks to use v1/2/3/4. But will take way longer. And will definitely break compatibility with existing saves. But we can also get tagging on perks. Which could help in combat too... maybe?