"True" Mutations, Mutagen tweaks.
Mutations now have a "true" variant, of which usually only PermTF mutations can access.

True mutations must be tagged as such under their Supers.
True Mutations can be checked with their v3 values, which will be set automatically, when a PermTF event triggers, and wipes all existing Mutations.
True Mutations can use the above value to call trueMutationBuffs under IMutationPerkType to build the buffs in a standard way.
True Mutations will also be indicated when you open the Database.
Note: This function is currently incomplete. It is missing which stats they will buff, as well as which Mutations qualify for taking in "True" variants.

True Mutations were designed due to the way some Mutations were instead planned to be acquirable naturally, so, that idea was buffed, and placed behind this moniker, to prevent Evangeline from effectively being entirely useless in Mutations functions.

-----
Mutations are now able to bypass Level requirements if given Mutagen instead of Gems for the tf cost.

This was due to players indicating the costs of gems vs the RNG and acquisition of the Mutagen to be unbalanced.
This will likely be further modified, to further incentivize getting Mutagens.