New Hellfire. Changes w/ Scale + Dmg
Touched up hellfire Damage and Scaling to bring it better in line with other abilities. Now Scales off of Lib and Str. It splits the damage between fire and lust depending on monsters corruption. It also scales off of PC corruption.

New PC corruption scaling for those willing to use it in combat.as Currently it requires a minimum of 25 corruption to get any additive benefit and a corruption stat above 55 to break even or buff it if you multiply base X corrupt scaling.